requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r requirements/base.in
certifi==2022.12.7
    # via requests
cffi==1.15.1
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   edx-django-utils
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via
    #   django-rest-swagger
    #   openapi-codec
coreschema==0.0.4
    # via coreapi
cryptography==39.0.1
    # via
    #   pyjwt
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==3.2.18
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
    #   django-cors-headers
    #   django-crum
    #   django-extensions
    #   djangorestframework
    #   drf-jwt
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
django-cors-headers==3.13.0
    # via -r requirements/base.in
django-crum==0.7.9
    # via edx-django-utils
django-extensions==3.2.1
    # via -r requirements/base.in
django-rest-swagger==2.2.0
    # via -r requirements/base.in
django-waffle==3.0.0
    # via
    #   -r requirements/base.in
    #   edx-django-utils
    #   edx-drf-extensions
djangorestframework==3.14.0
    # via
    #   -r requirements/base.in
    #   django-rest-swagger
    #   drf-jwt
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via edx-drf-extensions
edx-auth-backends==4.1.0
    # via -r requirements/base.in
edx-django-release-util==1.2.0
    # via -r requirements/base.in
edx-django-utils==5.2.0
    # via
    #   -r requirements/base.in
    #   edx-drf-extensions
    #   edx-rest-api-client
edx-drf-extensions==8.4.1
    # via -r requirements/base.in
edx-opaque-keys==2.3.0
    # via edx-drf-extensions
edx-rest-api-client==5.5.0
    # via -r requirements/base.in
future==0.18.3
    # via pyjwkest
idna==3.4
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
kombu==5.2.4
    # via celery
markupsafe==2.1.2
    # via jinja2
mysqlclient==2.1.1
    # via -r requirements/base.in
newrelic==8.7.0
    # via edx-django-utils
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
openapi-codec==1.3.2
    # via django-rest-swagger
pbr==5.11.1
    # via stevedore
prompt-toolkit==3.0.37
    # via click-repl
psutil==5.9.4
    # via edx-django-utils
pycparser==2.21
    # via cffi
pycryptodomex==3.17
    # via pyjwkest
pyjwkest==1.4.2
    # via edx-drf-extensions
pyjwt[crypto]==2.6.0
    # via
    #   drf-jwt
    #   edx-auth-backends
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   social-auth-core
pymongo==3.13.0
    # via edx-opaque-keys
pynacl==1.5.0
    # via edx-django-utils
python-dateutil==2.8.2
    # via edx-drf-extensions
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.7.1
    # via
    #   -r requirements/base.in
    #   celery
    #   django
    #   djangorestframework
pyyaml==6.0
    # via edx-django-release-util
redis==4.5.1
    # via -r requirements/base.in
requests==2.28.2
    # via
    #   coreapi
    #   edx-drf-extensions
    #   edx-rest-api-client
    #   pyjwkest
    #   requests-oauthlib
    #   slumber
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
semantic-version==2.10.0
    # via edx-drf-extensions
simplejson==3.18.3
    # via django-rest-swagger
six==1.16.0
    # via
    #   click-repl
    #   edx-auth-backends
    #   edx-django-release-util
    #   edx-drf-extensions
    #   pyjwkest
    #   python-dateutil
slumber==0.7.1
    # via edx-rest-api-client
social-auth-app-django==5.0.0
    # via edx-auth-backends
social-auth-core==4.3.0
    # via
    #   edx-auth-backends
    #   social-auth-app-django
sqlparse==0.4.3
    # via django
stevedore==5.0.0
    # via
    #   edx-django-utils
    #   edx-opaque-keys
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.14
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
